Separate NDJSON direct planning
